change the hover on the offer card, the pin is highlighted in orange
